pyproject.toml
[project]
name = "langchain-gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-agent>=0.1.1",
    "langchain-core>=1.2.8",
    "langchain-google-community[gmail]>=3.0.5",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]
